at index

+++ 0x17-doubly_linked_lists/5-get_dnodeint.c
@@ -30,7 +30,7 @@ return (res);
 dlistint_t *get_dnodeint_at_index(dlistint_t *head, unsigned int index)
 {
 dlistint_t *tmp = head;
-if (index >= dlistint_len(head) || head == NULL)
+if ((int)index >= (int)dlistint_len(head) || head == NULL)
 {
 return (NULL);
 }
